datasets/requirements.txt
pandas==2.2.2
paho-mqtt==2.1.0
